+++ Changelog.md
@@ -4,6 +4,7 @@
 ### New features
 * Use bootbox alert instead of pure JS for changing branches
 * Ability to set DO server size on server creation
+* Copy ip of server to clipboard on click
 
 ### Bug Fixes
 * Fix server error while showing html result for spec with zero specs